Unit tests for text editor menu
RISDEV-4432
