+++ frontend/test/components/input/textEditorMenu.spec.ts
@@ -1,39 +1,15 @@
 /* eslint-disable testing-library/no-node-access */
 import { userEvent } from "@testing-library/user-event"
-import { render, screen } from "@testing-library/vue"
+import { render, screen, waitFor } from "@testing-library/vue"
 import { flushPromises } from "@vue/test-utils"
-import { beforeEach } from "vitest"
 import { createRouter, createWebHistory } from "vue-router"
 import TextEditor from "@/components/input/TextEditor.vue"
+import { mockDocumentForProsemirror } from "~/components/input/prosemirror-document-mock"
 
-function getBoundingClientRect() {
-  const rec = {
-    x: 0,
-    y: 0,
-    bottom: 0,
-    height: 0,
-    left: 0,
-    right: 0,
-    top: 0,
-    width: 0,
-  }
-  return { ...rec, toJSON: () => rec }
-}
+mockDocumentForProsemirror()
 
-class FakeDOMRectList extends DOMRect {
-  item(index: any) {
-    return (this as any)[index]
-  }
-}
-
-document.elementFromPoint = () => null
-HTMLElement.prototype.getBoundingClientRect = getBoundingClientRect
-HTMLElement.prototype.getClientRects = () => new FakeDOMRectList() as any
-Range.prototype.getBoundingClientRect = getBoundingClientRect
-Range.prototype.getClientRects = () => new FakeDOMRectList() as any
-
-describe("text editor", async () => {
-  beforeEach(async () => {
+describe("text editor toolbar", async () => {
+  const renderComponent = async () => {
     userEvent.setup()
     render(TextEditor, {
       props: {
@@ -45,7 +21,7 @@ describe("text editor", async () => {
     })
 
     await flushPromises()
-  })
+  }
 
   global.ResizeObserver = require("resize-observer-polyfill")
   const router = createRouter({
@@ -64,54 +40,108 @@ describe("text editor", async () => {
     ],
   })
 
-  test("shift tab in text editor should focus first button in menu", async () => {
-    const editorField = screen.getByTestId("Test Editor Feld")
+  describe("keyboard navigation", () => {
+    test("shift tab in text editor should focus first button in menu", async () => {
+      await renderComponent()
+      const editorField = screen.getByTestId("Test Editor Feld")
 
-    await userEvent.click(editorField.firstElementChild!)
-    expect(editorField.firstElementChild).toHaveFocus()
-    await userEvent.tab({ shift: true })
-    const firstButton = screen.getByLabelText("fullview")
-    expect(firstButton).toHaveFocus()
-  })
+      await userEvent.click(editorField.firstElementChild!)
+      expect(editorField.firstElementChild).toHaveFocus()
+      await userEvent.tab({ shift: true })
+      const firstButton = screen.getByLabelText("fullview")
+      expect(firstButton).toHaveFocus()
+    })
 
-  test("arrow right should move focus to next button until last button", async () => {
-    const editorField = screen.getByTestId("Test Editor Feld")
+    test("arrow right should move focus to next button until last button", async () => {
+      await renderComponent()
+      const editorField = screen.getByTestId("Test Editor Feld")
 
-    await userEvent.click(editorField.firstElementChild!)
-    expect(editorField.firstElementChild).toHaveFocus()
-    await userEvent.tab({ shift: true })
-    const firstButton = screen.getByLabelText("fullview")
-    expect(firstButton).toHaveFocus()
+      await userEvent.click(editorField.firstElementChild!)
+      expect(editorField.firstElementChild).toHaveFocus()
+      await userEvent.tab({ shift: true })
+      const firstButton = screen.getByLabelText("fullview")
+      expect(firstButton).toHaveFocus()
 
-    await userEvent.keyboard("{ArrowRight}")
-    const secondButton = screen.getByLabelText("invisible-characters")
-    expect(secondButton).toHaveFocus()
+      await userEvent.keyboard("{ArrowRight}")
+      const secondButton = screen.getByLabelText("invisible-characters")
+      expect(secondButton).toHaveFocus()
 
-    // navigate to last button
-    await userEvent.keyboard("{ArrowRight>18/}")
-    const lastButton = screen.getByLabelText("redo")
-    expect(lastButton).toHaveFocus()
+      // navigate to last button (arrow right 18 times)
+      await userEvent.keyboard("{ArrowRight>18/}")
+      const lastButton = screen.getByLabelText("redo")
+      expect(lastButton).toHaveFocus()
 
-    // go one step further to the right --> should do nothing
-    await userEvent.keyboard("{ArrowRight}")
-    expect(lastButton).toHaveFocus()
+      // go one step further to the right --> should do nothing
+      await userEvent.keyboard("{ArrowRight}")
+      expect(lastButton).toHaveFocus()
 
-    // navigate to the left --> should go to previous button
-    await userEvent.keyboard("{ArrowLeft}")
-    const secondLastButton = screen.getByLabelText("undo")
-    expect(secondLastButton).toHaveFocus()
-  })
+      // navigate to the left --> should go to previous button
+      await userEvent.keyboard("{ArrowLeft}")
+      const secondLastButton = screen.getByLabelText("undo")
+      expect(secondLastButton).toHaveFocus()
+    })
 
-  test("arrow left should leave focus on first button", async () => {
-    const editorField = screen.getByTestId("Test Editor Feld")
+    test("arrow left should leave focus on first button", async () => {
+      await renderComponent()
+      const editorField = screen.getByTestId("Test Editor Feld")
 
-    await userEvent.click(editorField.firstElementChild!)
-    expect(editorField.firstElementChild).toHaveFocus()
-    await userEvent.tab({ shift: true })
-    const firstButton = screen.getByLabelText("fullview")
-    expect(firstButton).toHaveFocus()
+      await userEvent.click(editorField.firstElementChild!)
+      expect(editorField.firstElementChild).toHaveFocus()
+      await userEvent.tab({ shift: true })
+      const firstButton = screen.getByLabelText("fullview")
+      expect(firstButton).toHaveFocus()
 
-    await userEvent.keyboard("{ArrowLeft}")
-    expect(firstButton).toHaveFocus()
+      // When the first button is focused, ArrowLeft does not move focus
+      await userEvent.keyboard("{ArrowLeft>5}")
+      expect(firstButton).toHaveFocus()
+
+      // From the first button you can move immediately to the next one
+      await userEvent.keyboard("{ArrowRight}")
+      const secondButton = screen.getByLabelText("invisible-characters")
+      expect(secondButton).toHaveFocus()
+
+      await userEvent.keyboard("{ArrowLeft}")
+      expect(firstButton).toHaveFocus()
+    })
+
+    test("enter should jump back to the editor input", async () => {
+      await renderComponent()
+      const editorField = screen.getByTestId("Test Editor Feld")
+
+      await userEvent.click(editorField.firstElementChild!)
+      await userEvent.tab({ shift: true })
+      await userEvent.keyboard("{ArrowRight}")
+      await userEvent.keyboard("{ArrowRight}")
+      const thirdButton = screen.getByLabelText("bold")
+      expect(thirdButton).toHaveFocus()
+
+      // When clicking enter on a text edit button, the focus moves to the editor
+      await userEvent.keyboard("{Enter}")
+      await waitFor(() => expect(editorField.firstElementChild).toHaveFocus(), {
+        timeout: 100,
+      })
+    })
+
+    test("tab into the editor should skip the menu tool bar", async () => {
+      await renderComponent()
+      const editorField = screen.getByTestId("Test Editor Feld")
+
+      // Add external input field to be focused first
+      const inputField = editorField.ownerDocument.createElement("input")
+      editorField.ownerDocument.body.prepend(inputField)
+      await userEvent.click(inputField)
+
+      // Tab skips the menu toolbar and focuses the editor content directly
+      await userEvent.tab()
+      expect(editorField.firstElementChild).toHaveFocus()
+
+      // ProseMirror needs to be focused explicitly as it does not trigger button enabling otherwise in the unit test
+      await userEvent.click(editorField.firstElementChild!)
+
+      // Tab back focuses the toolbar buttons
+      await userEvent.tab({ shift: true })
+      const firstButton = screen.getByLabelText("fullview")
+      expect(firstButton).toHaveFocus()
+    })
   })
 })
